requirements.txt
google-generativeai
fastapi[standard]
uvicorn
websockets
jinja2
httpx
redis
pybit
sse-starlette
pandas
cryptography
openai
langchain-ollama
langchain-core
pydantic
scipy
numpy==1.26.4
torch==2.2.2+cpu --index-url https://download.pytorch.org/whl/cpu # CPU-only version for compatibility
gymnasium
stable-baselines3
tqdm
ta


google-generativeai
pandas_ta
setuptools<81
redis
plotly
kaleido

yfinance
mplfinance
matplotlib

dash
google_auth_oauthlib
ccxt

ntplib
django
importlib-metadata==4.11.4

transformers
accelerate
slowapi
requests
pytesseract
Pillow
python-binance
ultralytics
openai-whisper
pydub
youtube-transcript-api
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
